Wrote getversion() method
I wrote the getversion() method to find the lilypond version.
It tries to run `lilypond --version` and grab the output. If that
doesn't run, or cannot find a number in the output, it will ask the
user. If lilypond fails to run, it will also prompt the user to check
that the install is working properly

updated TODO to reflect update