Wrote getlanguage() method
getlanguage() in getlanguage.rb will ask for the user to input the
language intended for lilypond input, normalizes it and checks it
against a list of Lilypond's supported languages (manual in file).
If it passes it returns a language statement, if not the user is given
a list of supported languages and prompted again.

Also updated TODO and Readme to reflect completion and new goals.