Add some ensembles and add some validators
A new command has been added to the db script to add ensembles. Some
ensembles have been added. This brings up the need to support for modifying ensembles and
recommitting them in the main program.

Also, I discovered the prompt toolkit validator, which is much cleaner than
the cumbersome validation loops in place before. Most of the prompts have
been updated to support this

Signed-off-by: Rick Henry <fredericmhenry@gmail.com>