Fixed some bugs in the refactor and added some ensembles
Some more ensembles have been added to the default db and the cli is better,
some more things have been moved to interface/common.

Large ensembles can now be easily entered to the database, but there is no
mechanism for indicating groupings of instruments, perhaps that will go in
the template.

Signed-off-by: Rick Henry <fredericmhenry@gmail.com>